requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
colorama==0.4.6
Django==5.0.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
